Rewrite the library using data kind promotion
This is a complete rewrite of the library, which is now based on the
DataKinds language extension introduced in GHC 7.4, and is released
under the terms of the BSD3 license.

Moreover the design of the library is also changed: it provides a
standard MultiIndex and Tensor type, while custom multiindex and tensor
types can be defined and must be instance of IsMultiIndex and IsTensor
classes respectively. The Ordinal types are gone, replaced by
multiindices with one component.

Features previously defined in Data.Tensor.LineareAlgebra and a proper
documentation are still missing and will be reintroduced in future
commits.